Fixed detection of Courier, indent for code blocks
Courier New was detected by comparing font family of text to
DocumentApp.FontFamily.COURIER_NEW using ===, this apparently now
fails. Added a function to do comparison with ==.

Added detection of indentation, so that we consider indented lines to
be equivalent to lines starting with a tab, in terms of code blocks.