Using GitHub Markdown Tables
These also work on Bitbucket.  I'm not sure how you use this tool, but I use it to convert Technical documents from google drive to live in a codebase instead (Where devs are more likely to find them).  It may also be cool to check lengths of table cell contents to even out the spacing to be more legible.